HIVE-2133. DROP TABLE IF EXISTS should not fail if a view of that name exists (John Sichi via cws)
git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1097709 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/DDLTask.java
@@ -2873,11 +2873,18 @@ public class DDLTask extends Task<DDLWork> implements Serializable {
       if (tbl != null) {
         if (tbl.isView()) {
           if (!dropTbl.getExpectView()) {
+            if (dropTbl.getIfExists()) {
+              return 0;
+            }
             throw new HiveException("Cannot drop a view with DROP TABLE");
           }
         } else {
           if (dropTbl.getExpectView()) {
-            throw new HiveException("Cannot drop a base table with DROP VIEW");
+            if (dropTbl.getIfExists()) {
+              return 0;
+            }
+            throw new HiveException(
+              "Cannot drop a base table with DROP VIEW");
           }
         }
       }
